feat(apps/web): users can add place from list
